feat(requests/send): Add mechanism to ensure that a req be sent and succeeds only once
Summary:
This commit adds a new option to NylasAPIRequest api to allow users to
specify requests that should only ever be executed once, once they
succeed. This is done via saving a hash of some of the requests
properties to the filesystem.

Specifically regarding send, this ensures that a draft never be sent
twice from N1. We ran into this case once when the final state of the TaskQueue
wasn't saved to the database before the N1 app closed. In this case, a
PerformSendActionTask that was supposed to be dequeued, was restored to
the task queue again when N1 restarted, causing the draft to be sent
twice.

Test Plan: Manual

Reviewers: evan, bengotow

Reviewed By: bengotow

Differential Revision: https://phab.nylas.com/D3407