HACK HACK HACK Allow querying account tokens by email address too.
We've got a plethora of "unique" ids for the same account and ot looks
like a recent regression is preventing us from fetching an API token for
an account from the keychain. This diff changes the AccountStore to be
able to fetch data using an account's email address. Will figure how to
do things better after launch.

Test plan: Checked that it worked for a broken account.