[client-app] Fixup analytics store
Summary:
Given that we no longer report analytics or metrics in dev mode, it
doesn't make sense to include that property when reporting to mixpanel.

Otherwise, it might cause confusion to future readers of this code as to
why we are sending `inDevMode` data if we never reach that code path when we
are in dev mode.

If we want to restore this in the future, we can just revert this
commit.

Test Plan: manual

Reviewers: evan

Reviewed By: evan

Differential Revision: https://phab.nylas.com/D4027